server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyjwt==2.1.0
requests==2.26.0
mysql-connector-python==8.0.26
cryptography==3.4.8
argon2-cffi==21.1.0
